1 使用了nacos作为config，但是在使用config.info这样依然会报错，提示还是未发现此种占位符，注意文件是正确的

+++ cloudalibaba-consumer-nacos-order83/src/main/java/com/github/greekpanda/cloudalibaba2020/controller/OrderController.java
@@ -20,11 +20,12 @@ public class OrderController {
     @Resource
     private RestTemplate restTemplate;
 
+    //读取配置文件
     @Value("${service-url.nacos-user-service}")
     private String serverURL;
 
     @GetMapping(value = "/consumer/payment/nacos/{id}")
-    public String paymentInfo(@PathVariable("id") Integer id) {
+    public String paymentInfo(@PathVariable("id") Long id) {
         return restTemplate.getForObject(serverURL + "/payment/nacos/" + id, String.class);
     }
